Refactor shader code into separate modules and add some shader tests (#2664)
* add foreign key constraints to schema and as evolution #2533

* added the evolution #2533

* implement pr advice #2568

* corrected pr number in evolution #2568

* merge master, switch evolution numbers

* switch evolution numbers (not added in previous commit)

* fix inconsistencies of test db #2568

* fix annotation.csv and prepareTestDB #2568

* make the owner of a script non-optional #2612

* fix task assignment #2568

* correct annotations.csv and add error messages to Task finding #2568

* fix ObjectIdReads #2612

* fix script find all to only show scripts from your organization #2613

* remove cycles in sql schema and fix tests and global.scala

* add evolution for new isOrganizationTeam field

* temporarily ignore some build steps

* tmp: shader rendering test

* add pkg-config to dockerfile

* add glsify

* fix deleteAccessQ #2568

* rename evolutions and fix evolution for foreign key constraint #2568

* rename + fix evolutions

* ensure display is available when running shader tests

* implement shader module system + dependency resolving code; refactor shader code into seperate modules

* add test-driven syntax validation for shaders; refactor; fix tests

* enhance evolution to set organization team information #2568

* use headless-gl fork with prebuilt binaries

* update yarn.lock

* pin headless-gl version

* use newer headless-gl version to allow CI using prebuilt package

* revert unnecessary changes

* added Failure-messages to debug error

* allow refresh datasets for teammanagers #2592

* remove unnecessary comments and import

* fix datasets update query for zero-datasets case

* respect packing degree of segmentation layer in shader (#2693)

* ensure that only buckets for existing zoom steps are pulled

* annotate foreign-key sql evolution

* upgrade grpc and use new fossildb standard health check (#2637)

* WIP upgrade grpc

* in datastore health check, also check fossilDB. remove scalapb-json4s

* bump fossildb version to 0.1.10 (includes new docker login)

* fix datastore merge conflict

* CI: make curl fail on non-200 responses

* upgrade fossildb also in docker-compose

* make standalone-datastore work in docker (#2705)

* first changes to make datastore work in docker

* make this useful for everybody
