+++ .circleci/config.yml
@@ -94,7 +94,7 @@ jobs:
           name: Run webknossos-datastore smoke test
           command: |
             sleep 10
-            curl --retry 3 --max-time 15 -v http://localhost:9090/data/health
+            curl --fail --retry 3 --max-time 15 -v http://localhost:9090/data/health
       - run:
           name: Stop webknossos-datastore
           command: docker-compose down --volumes --remove-orphans
