+++ .flowconfig
@@ -13,6 +13,7 @@
 <PROJECT_ROOT>/public/.*
 <PROJECT_ROOT>/tools/.*
 <PROJECT_ROOT>/node_modules/webpack-cli/
+<PROJECT_ROOT>/node_modules/findup/
 
 [untyped]
 <PROJECT_ROOT>/node_modules/ava/
