+++ app/assets/javascripts/oxalis/model/bucket_data_handling/data_rendering_logic.js
@@ -41,7 +41,7 @@ export function validateMinimumRequirements(specs: GpuSpecs): void {
 
 export type DataTextureSizeAndCount = { textureSize: number, textureCount: number };
 
-function getPackingDegree(byteCount: number) {
+export function getPackingDegree(byteCount: number) {
   // If the layer only holds one byte per voxel, we can pack 4 bytes using rgba channels
   return byteCount === 1 ? 4 : 1;
 }
