+++ app/models/team/Organization.scala
@@ -20,7 +20,6 @@ import scala.concurrent.Future
 
 case class OrganizationSQL(
                             _id: ObjectId,
-                            _organizationTeam: ObjectId,
                             name: String,
                             additionalInformation: String,
                             created: Long = System.currentTimeMillis(),
@@ -33,9 +32,8 @@ object OrganizationSQL {
     Fox.successful(
       OrganizationSQL(
         ObjectId.fromBsonId(o._id),
-        ObjectId.fromBsonId(o._organizationTeam),
         o.name,
-        additionalInformation = o.additionalInformation
+        o.additionalInformation
       )
     )
   }
@@ -48,11 +46,11 @@ object OrganizationSQLDAO extends SQLDAO[OrganizationSQL, OrganizationsRow, Orga
 
   def isDeletedColumn(x: Organizations): Rep[Boolean] = x.isdeleted
 
+
   def parse(r: OrganizationsRow): Fox[OrganizationSQL] =
     Fox.successful(
       OrganizationSQL(
         ObjectId(r._Id),
-        ObjectId(r._Organizationteam),
         r.name,
         r.additionalinformation,
         r.created.getTime,
@@ -75,10 +73,17 @@ object OrganizationSQLDAO extends SQLDAO[OrganizationSQL, OrganizationsRow, Orga
   def insertOne(o: OrganizationSQL)(implicit ctx: DBAccessContext): Fox[Unit] =
     for {
       r <- run(
-        sqlu"""insert into webknossos.organizations(_id, _organizationTeam, name, additionalInformation, created, isDeleted)
-                  values(${o._id.id}, ${o._organizationTeam}, ${o.name}, ${o.additionalInformation}, ${new java.sql.Timestamp(o.created)}, ${o.isDeleted})
+
+        sqlu"""insert into webknossos.organizations(_id, name, additionalInformation, created, isDeleted)
+                  values(${o._id.id}, ${o.name}, ${o.additionalInformation}, ${new java.sql.Timestamp(o.created)}, ${o.isDeleted})
             """)
     } yield ()
+
+  def findOrganizationTeam(o: ObjectId) =
+    for{
+      r <- run(sql"select _id from webknossos.organizationTeams where _organization = ${o.id}".as[String])
+      parsed <- BSONObjectID.parse(r.head).toOption.toFox ?~> Messages("sql.invalidBSONObjectId")
+    } yield ObjectId.fromBsonId(parsed)
 
 }
 
@@ -101,7 +106,8 @@ object Organization extends FoxImplicits {
   def fromOrganizationSQL(o: OrganizationSQL)(implicit ctx: DBAccessContext) = {
     for {
       idBson <- o._id.toBSONObjectId.toFox ?~> Messages("sql.invalidBSONObjectId")
-      organizationTeamIdBson <- o._organizationTeam.toBSONObjectId.toFox ?~> Messages("sql.invalidBSONObjectId")
+      organizationTeamId <- OrganizationSQLDAO.findOrganizationTeam(o._id).toFox
+      organizationTeamIdBson <- organizationTeamId.toBSONObjectId.toFox ?~> Messages("sql.invalidBSONObjectId")
       teams <- TeamSQLDAO.findAllByOrganization(o._id)
       teamBsonIds <- Fox.combined(teams.map(_._id.toBSONObjectId.toFox))
     } yield {
