+++ app/models/user/time/TimeSpanDAO.scala
@@ -35,8 +35,8 @@ object TimeSpanDAO {
 
   def update(timeSpan: TimeSpan)(implicit ctx: DBAccessContext) = {
     for {
-      timeSpanSQL <- TimeSpanSQL.fromTimeSpan(timeSpan)
-      _ <- TimeSpanSQLDAO.updateOne(timeSpanSQL)
+      timeSpanSQL <- TimeSpanSQL.fromTimeSpan(timeSpan) ?~> "FAILED: TimeSpanSQL.fromTimeSpan"
+      _ <- TimeSpanSQLDAO.updateOne(timeSpanSQL)  ?~> "FAILED: TimeSpanSQLDAO.updateOne"
     } yield ()
   }
 }
