+++ app/utils/SQLHelpers.scala
@@ -145,7 +145,7 @@ trait SecuredSQLDAO extends SimpleSQLDAO {
     if (ctx.globalAccess) Fox.successful(())
     else {
       for {
-        userId <- userIdFromCtx
+        userId <- userIdFromCtx ?~> "FAILED: userIdFromCtx"
         resultList <- run(sql"select _id from #${existingCollectionName} where _id = ${id.id} and #${updateAccessQ(userId)}".as[String]) ?~> "Failed to check write access. Does the object exist?"
         _ <- resultList.headOption.toFox ?~> "Access denied."
       } yield ()
