+++ docker-compose.yml
@@ -17,6 +17,9 @@ services:
       - -Dhttp.address=0.0.0.0
       - -Ddatastore.fossildb.address=fossildb
       - -Dpostgres.url=jdbc:postgresql://postgres/webknossos
+      # the following lines disable the integrated datastore:
+      # - -Dplay.modules.enabled-="com.scalableminds.webknossos.datastore.DataStoreModule"
+      # - -Ddatastore.enabled=false
     environment:
       - POSTGRES_URL=jdbc:postgresql://postgres/webknossos
 
@@ -25,7 +28,22 @@ services:
     image: scalableminds/webknossos-datastore:${DOCKER_TAG:-master}
     ports:
       - "9090:9090"
-
+    entrypoint: bin/webknossos-datastore
+    volumes:
+      - ./binaryData:/srv/webknossos-datastore/binaryData
+    command:
+      - -J-Xmx20G
+      - -J-Xms1G
+      - -Dconfig.file=conf/standalone-datastore.conf
+      - -Dlogger.file=conf/logback-docker.xml
+      - -Dlogback.configurationFile=conf/logback-docker.xml
+      - -Dhttp.port=9090
+      - -Dhttp.address=0.0.0.0
+      - -Dhttp.uri=http://webknossos-datastore:9090
+      - -Ddatastore.fossildb.address=fossildb
+      - -Ddatastore.oxalis.uri=webknossos:9000
+    links:
+      - fossildb
 
   base:
     image: scalableminds/sbt:sbt-0.13.15_mongo-3.4.10_node-9.x_jdk-8
@@ -113,8 +131,16 @@ services:
     volumes:
       - "./pg/dev-db:/var/lib/postgresql/data/"
 
+  psql:
+    extends: postgres
+    command: psql -h postgres -U postgres webknossos
+    links:
+      - postgres
+    environment:
+      PGPASSWORD: postgres
+
   fossildb:
-    image: scalableminds/fossildb:master__272
+    image: scalableminds/fossildb:master__337
     command:
       - fossildb
       - -c
