+++ fossildb/version
@@ -1,1 +1,1 @@
-0.1.7
+0.1.10
