+++ package.json
@@ -41,6 +41,7 @@
     "file-loader": "^1.1.5",
     "flow-bin": "^0.71.0",
     "flow-coverage-report": "^0.5.0",
+    "gl": "scalableminds/headless-gl#4.0.5",
     "himalaya": "^0.2.12",
     "isomorphic-fetch": "^2.2.0",
     "jsdom": "^11.5.1",
@@ -50,6 +51,7 @@
     "lint-staged": "^7.1.2",
     "mock-require": "^1.2.1",
     "node-fetch": "^1.7.2",
+    "pngjs": "latest",
     "prettier": "1.11.1",
     "proto-loader6": "^0.4.0",
     "react-test-renderer": "^16.2.0",
@@ -62,11 +64,11 @@
     "webpack-cli": "^2.1.3"
   },
   "scripts": {
-    "start": "./sbt run -Dconfig.file=conf/application.conf -jvm-debug 5005",
+    "start": "./sbt run -jvm-debug 5005",
     "build": "node_modules/.bin/webpack --env.production",
     "build-watch": "node_modules/.bin/webpack -w",
     "test": "tools/test.sh prepare -q && tools/test.sh test --timeout=30s",
-    "test-verbose": "tools/test.sh prepare -q && tools/test.sh test --timeout=60s --verbose",
+    "test-verbose": "tools/test.sh prepare -q && xvfb-run -s '-ac -screen 0 1280x1024x24' tools/test.sh test --timeout=60s --verbose",
     "test-only": "tools/test.sh test --timeout=30s",
     "test-watch": "tools/test.sh test --watch",
     "test-debug": "tools/test.sh test --debug",
@@ -107,6 +109,7 @@
     "es6-promise": "^3.0.2",
     "file-saver": "^1.3.3",
     "font-awesome": "^4.5.0",
+    "glsl-parser": "git+https://github.com/anvaka/glslx.git#glsl-parser",
     "hammerjs": "^2.0.8",
     "history": "^4.7.2",
     "immutability-helper": "^2.6.6",
