+++ test/db/annotations.csv
@@ -1,6 +1,5 @@
 _id,_dataSet,_task,_team,_user,tracing_id,tracing_typ,description,isPublic,name,state,statistics,tags,tracingTime,typ,created,modified,isDeleted
-'570b9ff12a7c0e980056fe8f','570b9f4e4bb848d0885ee711','570b9ff02a7c0e960056fe8b','570b9f4b2a7c0e3b008da6ec','570b9f4d2a7c0e4d008da6ef','4ab765c1-1089-4dc8-98c7-7cce4879f7e9','skeleton','',f,'','Active','{}','{}',,'TracingBase','1970-01-01T00:00:00.000Z','1970-01-01T00:00:00.000Z',t
+'570b9ff12a7c0e980056fe8f','570b9f4e4bb848d0885ee711','581367a82faeb37a008a5352','570b9f4b2a7c0e3b008da6ec','570b9f4d2a7c0e4d008da6ef','4ab765c1-1089-4dc8-98c7-7cce4879f7e9','skeleton','',f,'','Active','{}','{}',,'TracingBase','1970-01-01T00:00:00.000Z','1970-01-01T00:00:00.000Z',f
 '570ba0092a7c0e980056fe9b','570b9f4e4bb848d0885ee711',,'570b9f4b2a7c0e3b008da6ec','570b9f4d2a7c0e4d008da6ef','ae417175-f7bb-4a34-8187-d9c3b50143ae','skeleton','',f,'','Active','{"treeCount":2,"nodeCount":28967,"edgeCount":28965,"branchPointCount":0}','{}',,'Explorational','1970-01-01T00:00:00.000Z','1970-01-01T00:00:00.000Z',f
 '58135c192faeb34c0081c05c','570b9f4e4bb848d0885ee711','58135c192faeb34c0081c058','570b9f4b2a7c0e3b008da6ec','570b9f4d2a7c0e4d008da6ef','66220a3a-9715-4792-bed4-6fda7ca8193f','skeleton','',f,'','Active','{}','{}',,'TracingBase','1970-01-01T00:00:00.000Z','1970-01-01T00:00:00.000Z',f
-'58135c402faeb34e0081c068','570b9f4e4bb848d0885ee711','58135c192faeb34c0081c058','570b9f4b2a7c0e3b008da6ec','570b9f4d2a7c0e4d008da6ef','07fd0bd7-f7a6-46d1-aaa8-58a381da4d50','skeleton','',f,'','Active','{}','{}',,'Task','1970-01-01T00:00:00.000Z','1970-01-01T00:00:00.000Z',f
-'581367a82faeb37a008a5356','570b9f4e4bb848d0885ee711','581367a82faeb37a008a5352','570b9f4b2a7c0e3b008da6ec','570b9f4d2a7c0e4d008da6ef','d701af7d-c8a7-44bd-86b5-6f44ab407e75','skeleton','',f,'','Active','{}','{}',,'TracingBase','1970-01-01T00:00:00.000Z','1970-01-01T00:00:00.000Z',f
+'58135c402faeb34e0081c068','570b9f4e4bb848d0885ee711','581367a82faeb37a008a5352','570b9f4b2a7c0e3b008da6ec','570b9f4d2a7c0e4d008da6ef','07fd0bd7-f7a6-46d1-aaa8-58a381da4d50','skeleton','',f,'','Active','{}','{}',,'Task','1970-01-01T00:00:00.000Z','1970-01-01T00:00:00.000Z',f
